System.Win32.Async.ErrCode

Description

ErrCodes which can appear when working with Async

Documentation

eRROR_ABANDONED_WAIT_0 ∷ ErrCode

This error is thrown by GetQueuedCompletionsStatus when I/O completion port is closed.

735

eRROR_HANDLE_EOF ∷ ErrCode

Reached the end of file.

38

eRROR_INVALID_HANDLE ∷ ErrCode

The HANDLE is invalid.

6

eRROR_NO_DATA ∷ ErrCode

eRROR_OPERATION_ABORTED ∷ ErrCode

Error returned by GetQueuedCompletionsStatus when an I/O operation was cancelled with CancelIoEx.

995

eRROR_IO_INCOMPLETE ∷ ErrCode

eRROR_IO_PENDING ∷ ErrCode

For async operations, getLastError reports ERROR_IO_PENDING. It is not an error per se, but used by the os to notify that the i/o will be completed asynchronously. For the scope of this library it means that the IOManager thread will be notified when the operation completes.

997

eRROR_PIPE_CONNECTED ∷ ErrCode

eRROR_PIPE_LISTENING ∷ ErrCode

eRROR_NOT_FOUND ∷ ErrCode

type WSAErrCode = Int

wSA_IO_PENDING ∷ WSAErrCode

wsaGetLastError ∷ IO WSAErrCode

wsaFailWith ∷ String → WSAErrCode → IO a