System.Win32.Async.IOData

Description

This module is provides an internal mechanism of doing asynchronous IO.

Documentation

withIOCPData

Arguments

∷ ∀ a (asyncType ∷ AsyncType). Storable (IOData asyncType)

⇒ String
error tag

→ FD asyncType
file descriptor used for the syscall

→ (Ptr (OverlappedType asyncType) → MVar (Either ErrCode Int) → IO (Result asyncType a))
continuation, executed with unmasked async exceptions which executes a syscall. It receives LPOVERLAPPED allocated on process heap and an empty MVar which will be filled by dequeueCompletionPackets (IOManager thread), when the async IO terminates.

Note that the MVar holds 'ErrorCode Async', even if 'asyncType ~ WsaAsync', this is because the GetQueuedCompletionStatus and thus dequeueCompletionPackets can handle both `windows.h` and `winsock2.h` apis.

→ IO a

Allocate IOCPData on process heap. If the continuation does not succeed and getLastError is not ERROR_IO_PENDING it might deallocate IOCPData otherwise the responsibility of deallocation is in the dequeueCompletionPackets thread.

It is executed with async exception masked to ensure that ioDataPtr is filled in with IOCPData before it is passed to the continuation.

data AsyncType

Constructors

Async

WsaAsync

data AsyncSing (asyncType ∷ AsyncType) where

A singleton for AsyncType.

Constructors

AsyncSing ∷ AsyncSing Async

WsaAsyncSing ∷ AsyncSing WsaAsync

data IOData (asyncType ∷ AsyncType)

Data associated with each async call. It is represented as the IODATA struct.

When an async call is initialised, e.g. ReadFile, the windows api takes a pointer to OVERLAPPED or WSAOVERLAPPED in case of winsock2. When it completes GetQueuedCompletionStatus returns the OVERLAPPED (or WSAOVERLAPPED) in the same memory location. Since we keep OVERLAPPED (WSAOVERLAPPED) as a first member of _IODATA (_WSAIODATA) struct, we can cast a pointer to OVERLAPPED (WSAOVERLAPPED) to recover a pointer to _IODATA ('_WSAIODATA)' and thus recover 'IOData asyncType', including the stable pointer.

The dequeueCompletionPackets (which calls GetQueuedCompletionStatus fills the MVar with the result returned by GetQueuedCompletionStatus (or GetLastError).

Outside of this module the type synonim IOCPData, AsyncIOCPData or WsaIOCPData should be used.

type AsyncIOCPData = IOData Async

AsyncIOCPData type synonim guarantees that dequeueCompletionPackets is using the same dataa type as withIOCPData in this module, i.e. submitted overlapped requests matches with the notifications pulled by dequeueCompletionPackets.

castOverlappedPtr ∷ Ptr (OverlappedType asyncType) → Ptr (IOData asyncType)

We can safely cast a pointer to OverlappedType asyncType to a pointer to 'IOData asyncType. This is because IOData asyncType is represented by a struct IODATA (or WSAIODATA), which first member is OVERLAPPED (WSAOVERLAPPED respectively). See the Storable instance for IOData below.

type WsaAsyncIOCPData = IOData WsaAsync

iodOverlappedPtr ∷ AsyncSing asyncType → Ptr (IOData asyncType) → Ptr (OverlappedType asyncType)

Access the OVERLAPPED member of IODATA struct, in C:

 ioData->iodOverlapped

iodDataPtr ∷ AsyncSing asyncType → Ptr (IOData asyncType) → Ptr (StablePtr (MVar (f Int)))

Access iodData member of IODATA or WSAIODATA struct.

data ErrorCode asyncType where

Error codes: either ErrCode or WsaErrCode. For 'asyncType ~ WsaAsync' we allow to use both. This is because the dequeueCompletionPackets rightfully only reports ErrCodes.

Constructors

ErrorCode ∷ ∀ (any ∷ AsyncType). !ErrCode → ErrorCode any

WsaErrorCode ∷ !WSAErrCode → ErrorCode WsaAsync

data FD asyncType where

File descriptors.

Constructors

FDHandle ∷ HANDLE → FD Async

FDSocket ∷ SOCKET → FD WsaAsync

data Result asyncType a

Result of an async i/o action. It maybe either asynchronous or synchronous result or error.

One has to be really careful with passing True value to ResultSync or ErrorSync - it can result in a race between two threads deallocating the same memory, subsequently os killing the process (often silently).

Constructors

ResultAsync !a
Asynchronous result, as returned by the iocp thread.

ErrorAsync !(ErrorCode asyncType)
Asynchronous error, as returned by the iocp thread.

ResultSync !a !Bool
Synchronous result; the if the Bool value is true, ioDataPtr will be deallocated.

ErrorSync !(ErrorCode asyncType) !Bool
Synchronous error; the if the Bool value is true, ioDataPtr will be deallocated.