System.Win32.Async.IOManager

IOManager

data IOCompletionPort

New type wrapper which holds HANDLE of the I/O completion port. https://docs.microsoft.com/en-us/windows/win32/fileio/createiocompletionport

associateWithIOCompletionPort ∷ Either HANDLE Socket → IOCompletionPort → IO ()

Associate with I/O completion port. This can be used multiple times on a file descriptor.

withIOManager ∷ (IOCompletionPort → IO r) → IO r

Create an IOCompletionPort and start a I/O manager thread (bound thread). The IOManager thread will only run for the duration of the callback.

The IOManager will throw IOException back to the starting thread (which should be the main application thread). There's no point of running the application if the dequeueCompletionPackets died, which most likely indicates a bug in itself.

TODO: add a tracer which logs when dequeueCompletionPackets errors

data IOManagerError

A newtype wrapper for IOErrors whihc are catched in the IOManager thread and are re-thrown in the main application thread.

Low level IOCP interface

dequeueCompletionPackets

Arguments

∷ IOCompletionPort
handle of a completion port

→ IO ()

IO manager which handles completions of IO operations. It should run on a bound thread. It dereferences the stable pointer which was allocated by readAsync or writeAsync, and puts the results of the operation in the MVar which it points to. This wakes up the waitForCompletion function.

To safely shutdown the I/O manager, it's enough to close the port's HANDLE. This should be done after all the associated handles are closed.

The underlying c code is using a fixed completion key which is not exposed on the Haskell side.

createIOCompletionPort

Arguments

∷ DWORD
number of concurrent threads

→ IO IOCompletionPort

Windows documentation: https://docs.microsoft.com/en-us/windows/win32/fileio/createiocompletionport

closeIOCompletionPort ∷ IOCompletionPort → IO ()