System.Win32.Async.Internal

Description

This module exposed internals of IOManager for testing purposes, use IOManager instead.

Documentation

createIOCompletionPort

Arguments

∷ DWORD
number of concurrent threads

→ IO IOCompletionPort

Windows documentation: https://docs.microsoft.com/en-us/windows/win32/fileio/createiocompletionport

closeIOCompletionPort ∷ IOCompletionPort → IO ()

dequeueCompletionPackets

Arguments

∷ IOCompletionPort
handle of a completion port

→ IO ()

IO manager which handles completions of IO operations. It should run on a bound thread. It dereferences the stable pointer which was allocated by readAsync or writeAsync, and puts the results of the operation in the MVar which it points to. This wakes up the waitForCompletion function.

To safely shutdown the I/O manager, it's enough to close the port's HANDLE. This should be done after all the associated handles are closed.

The underlying c code is using a fixed completion key which is not exposed on the Haskell side.