System.Win32.Async.Socket.ByteString

Documentation

send ∷ Socket → ByteString → IO Int

Send a ByteString over a socket, which must be in a connected state, and must be associated with an IO completion port via associateWithIOManager. Returns number of bytes sent.

sendAll ∷ Socket → ByteString → IO ()

Like send but makes sure that all bytes are actually sent.

sendTo ∷ Socket → ByteString → SockAddr → IO Int

sendAllTo ∷ Socket → ByteString → SockAddr → IO ()

recv ∷ Socket → Int → IO ByteString

Receive a ByteString from a socket, which must be in a connected state, and must be associated with an IO completion port via associateWithIOManager. It may return less bytes than requested.

recvFrom

Arguments

∷ Socket
Socket

→ Int
Maximum number of bytes to receive

→ IO (ByteString, SockAddr)
Data received and sender address

Receive a ByteString from a socket together with address from which it came. The socket must be bound to a local address (see recvBufFrom), and must be associated with IO completion port via associateWithIOManager