System.Win32.Async.Socket.ByteString.Lazy

Documentation

send ∷ Socket → ByteString → IO Int64

Sending each chunk using vectored I/O. In one system call one can transmit at most 1024 chunks, it is safe to transmit less than 4194304 bytes.

The socket must be in a connected state, and must be associated with an IO completion port via associateWithIOManager.

sendAll ∷ Socket → ByteString → IO ()

recv ∷ Socket → Int → IO ByteString

Receive bytes from a socket, which must be in a connected state, and must be associated with an IO completion port via associateWithIOManager. It can return less bytes than requested.