Test.Generators

Documentation

tests ∷ TestTree

data NonEmptyBS

Small non-empty ByteStrings.

Constructors

NonEmptyBS

Fields

getNonEmptyBS ∷ ByteString

prop_NonEmptyBS ∷ NonEmptyBS → Bool

prop_shrink_NonEmptyBS ∷ NonEmptyBS → Bool

data LargeNonEmptyBS

Large non-empty ByteStrings, up to 2.5MB.

Constructors

LargeNonEmptyBS

Fields

getLargeNonEmptyBS ∷ ByteString

getSize ∷ Word

arbitrary size which is less than length of the bytestring; Useful for setting buffer size of a named pipe