System.Win32.Async.Socket

Documentation

sendBuf ∷ Socket → Ptr Word8 → Int → IO Int

sendBufTo

Arguments

∷ SocketAddress sa

⇒ Socket
Socket

→ Ptr Word8
data to send

→ Int
size of the data

→ sa
address to send to

→ IO Int

recvBuf ∷ Socket → Ptr Word8 → Int → IO Int

recvBufFrom ∷ SocketAddress sa ⇒ Socket → Ptr Word8 → Int → IO (Int, sa)

connect ∷ Socket → SockAddr → IO ()

Unfortunatelly connect using interruptible ffi is not interruptible. Instead we run the connect in a dedicated thread and block on an MVar.

accept ∷ Socket → IO (Socket, SockAddr)

This is a thin wrapper around accept. It's possible to killThread which runs the accept. It will leave a stranded thread, but closing a socket terminates accept call, and thus there's not resource leak.

TODO: other possible approaches: * use WSAEventSelect but it needs further investiation (unfortunatelly WSAAccept is not part of IOCP); or * use AcceptEx.